move tal/mni/mm/vox tests
